apps/crawler/requirements.txt
# Nodes Map Crawler - Python Dependencies

# Async networking
asyncio-dgram>=2.2.0
aiohttp>=3.11.0

# Database
supabase>=2.10.0
psycopg2-binary>=2.9.10

# GeoIP
maxminddb>=2.6.0
geoip2>=4.8.0

# Utilities
python-dotenv>=1.0.1
pyyaml>=6.0.2
structlog>=24.4.0

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
black>=24.10.0
ruff>=0.8.0
mypy>=1.13.0
